Create Profitable Apps profiles on Android and Apple Markets.ipynb
This project will be analyzing the profiles of dataset of
mobile phone apps that are currently available on both the
Apple Store and Google Play markets.

The aim is to use the conclusion of the analysis to predict
and decide on developing future apps that will be profitable on
both platforms.